:truck: moved navbar to everywhere
